fix: change virtualID type to uint8

+++ device/app/client.go
@@ -74,7 +74,7 @@ func (gclt *grpcClt) CreateSendTxn(
 		}
 		recvHandler(resp.Success, &Device{
 			Mac:       resp.Device.Mac,
-			VirtualID: resp.Device.VirtualID,
+			VirtualID: uint8(resp.Device.VirtualID),
 			Model:     resp.Device.Model,
 		}, resp.Error)
 	}
@@ -94,7 +94,7 @@ func (gclt *grpcClt) ModifySendTxn(
 				Act: pb.EditType_Add,
 				Device: &pb.Device{
 					Mac:       act.Device.Mac,
-					VirtualID: act.Device.VirtualID,
+					VirtualID: uint32(act.Device.VirtualID),
 					Model:     act.Device.Model,
 				},
 			})
@@ -275,7 +275,7 @@ func (gclt *grpcClt) GetDevicesByEquips(
 		for _, d := range resp.Devices {
 			mydevices = append(mydevices, &Device{
 				Mac:       d.MacAddress,
-				VirtualID: d.VirtualID,
+				VirtualID: uint8(d.VirtualID),
 				Model:     d.Model,
 			})
 		}
